+++ test/cbmc/proofs/MQTTPropGet_AssignedClientId/MQTTPropGet_AssignedClientId_harness.c
@@ -33,8 +33,8 @@ void harness()
 {
     MQTTPropBuilder_t * propBuffer;
     char ** pClientId;
-    uint16_t * clientIdLength;
-    uint32_t currentIndex;
+    size_t * clientIdLength;
+    size_t currentIndex;
 
     propBuffer = allocateMqttPropBuilder( NULL );
     __CPROVER_assume( isValidMqttPropBuilder( propBuffer ) );
@@ -47,7 +47,7 @@ void harness()
     }
 
     pClientId = malloc( sizeof( char * ) );
-    clientIdLength = malloc( sizeof( uint16_t ) );
+    clientIdLength = malloc( sizeof( size_t ) );
 
     MQTTPropGet_AssignedClientId( propBuffer, &currentIndex, pClientId, clientIdLength );
 }
