+++ test/cbmc/proofs/MQTT_SerializeUnsubscribe/MQTT_SerializeUnsubscribe_harness.c
@@ -33,14 +33,14 @@ void harness()
 {
     MQTTSubscribeInfo_t * pSubscriptionList;
     size_t subscriptionCount;
-    size_t remainingLength;
+    uint32_t remainingLength;
     uint16_t packetId;
     MQTTPropBuilder_t * pUnsubscribeProperties;
     uint32_t maxPacketSize;
 
     /* This variable is not used but is needed for MQTT_GetUnsubscribePacketSize()
      * to verify the pSubscriptionList. */
-    size_t packetSize;
+    uint32_t packetSize;
     MQTTFixedBuffer_t * pFixedBuffer;
     MQTTStatus_t status = MQTTSuccess;
